+++ client/components/Windows/Help.vue
@@ -16,15 +16,17 @@
 			</h2>
 
 			<div class="about">
-				<p
+				<div
 					v-if="store.state.serverConfiguration?.isUpdateAvailable"
-					class="update-available"
+					class="update-banner"
+					role="status"
 				>
-					A newer version of {{ appName }} has been downloaded and is ready to use.
-					<button type="button" class="btn" @click="reloadForUpdate">
-						Reload to update
+					<span class="update-banner-icon" aria-hidden="true"></span>
+					<span class="update-banner-text">New version available</span>
+					<button type="button" class="update-banner-reload" @click="reloadForUpdate">
+						Reload
 					</button>
-				</p>
+				</div>
 
 				<template v-if="pastRelease">
 					<p>